Make the buttons clickable, move the font out of the components folder, move basicPad to calcPads folder
