+++ src/index.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family: Tomorrow;
-    src: url("components/font/Tomorrow-Regular.ttf");
+    src: url("font/Tomorrow-Regular.ttf");
 }
 
 body {
